requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
openai
pydantic-settings
psycopg2-binary